Extensive changes to support summary reports which summarize the reports of subcomponents. Introduce new s.* expressions in JEXL
