+++ src/main/java/org/lsst/camera/portal/queries/SummaryUtils.java
@@ -53,7 +53,17 @@ public class SummaryUtils {
         }
         return result;
     }
-    public static Map<String, Map<String, List<Object>>> getReportValues2(HttpSession session, Integer actParentId, Integer reportId, String component) throws SQLException, ServletException, JspException {
+
+    public static Map<String, Map<String, Map<String, List<Object>>>> getReportValuesForSubcomponents(HttpSession session, Integer actParentId, Integer reportId, List<String> components) throws SQLException, ServletException, JspException {
+        Map<String, Map<String, Map<String, List<Object>>>> results = new LinkedHashMap<>();
+        // FIXME: We should not create independent database connectiosn for each iteration
+        for (String component : components) {
+            results.put(component, getReportValuesForSubcomponent(session, actParentId, reportId, component));
+        }
+        return results;
+    }
+
+    public static Map<String, Map<String, List<Object>>> getReportValuesForSubcomponent(HttpSession session, Integer actParentId, Integer reportId, String component) throws SQLException, ServletException, JspException {
         Map<String, Map<String, List<Object>>> result = new LinkedHashMap<>(); // orders the elements in the same order they're processed instead of random order.
 
         try (Connection c = ConnectionManager.getConnection(ModeSwitcherFilter.getVariable(session, "reportDisplayDb"))) {
@@ -85,5 +95,22 @@ public class SummaryUtils {
             }
         }
         return result;
+    }
+
+    public static Specifications getSpecifications(HttpSession session, Integer reportId) throws JspException, SQLException {
+        Specifications specs = new Specifications();
+        try (Connection c = ConnectionManager.getConnection(ModeSwitcherFilter.getVariable(session, "reportDisplayDb"))) {
+            PreparedStatement stmt = c.prepareStatement("select specid, jexl_value, jexl_status, jexl_measurement from report_specs where report=?");
+            stmt.setInt(1, reportId);
+            ResultSet r = stmt.executeQuery();
+            while (r.next()) {
+                String specId = r.getString("specid");
+                String jexl_value = r.getString("jexl_value");
+                String jexl_status = r.getString("jexl_status");
+                String jexl_measurement = r.getString("jexl_measurement");
+                specs.addSpec(specId,jexl_value,jexl_status,jexl_measurement);
+            }
+        }
+        return specs;
     }
 }
